# openapi/web-v1.json
{
  "components": {
    "schemas": {
...
      "CreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "content_mode": {
            "enum": [
              "pdf_preferred",
              "abstract_only"
            ],
            "title": "Content Mode",
            "type": "string"
          },
          "paper_limit": {
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Paper Limit",
            "type": "integer"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalRequest"
          }
        },
        "required": [
          "question",
          "paper_limit",
          "content_mode",
          "retrieval"
        ],
        "title": "CreateRunRequest",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "properties": {
          "artifact_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Run Id"
          },
          "available_artifacts": {
            "items": {
              "enum": [
                "papers.json",
                "documents.json",
                "evidence.json",
                "analyses.json",
                "report.json",
                "report.md",
                "run_manifest.json",
                "logs.jsonl"
              ],
              "type": "string"
            },
            "title": "Available Artifacts",
            "type": "array"
          },
          "content_mode": {
            "enum": [
              "pdf_preferred",
              "abstract_only"
            ],
            "title": "Content Mode",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "demo": {
            "title": "Demo",
            "type": "boolean"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "has_report": {
            "title": "Has Report",
            "type": "boolean"
          },
          "id": {
            "format": "uuid4",
            "title": "Id",
            "type": "string"
          },
          "manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ManifestProjection"
              },
              {
                "type": "null"
              }
            ]
          },
          "origin": {
            "enum": [
              "live",
              "bundled_demo"
            ],
            "title": "Origin",
            "type": "string"
          },
          "paper_limit": {
            "title": "Paper Limit",
            "type": "integer"
          },
          "phase": {
            "enum": [
              "queued",
              "initializing",
              "search",
              "acquisition",
              "chunking",
              "retrieval",
              "analysis",
              "synthesis",
              "citation_check",
              "publishing",
              "terminal"
            ],
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/RunProgress"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalRequest"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "completed_with_degradation",
              "failed",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "artifact_run_id",
          "origin",
          "status",
          "phase",
          "question",
          "paper_limit",
          "content_mode",
          "retrieval",
          "progress",
          "created_at",
          "started_at",
          "finished_at",
          "has_report",
          "demo",
          "manifest",
          "available_artifacts"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "RunList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "RunList",
        "type": "object"
      },
...
      "RunSummary": {
        "additionalProperties": false,
        "properties": {
          "artifact_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifact Run Id"
          },
          "content_mode": {
            "enum": [
              "pdf_preferred",
              "abstract_only"
            ],
            "title": "Content Mode",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "demo": {
            "title": "Demo",
            "type": "boolean"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "has_report": {
            "title": "Has Report",
            "type": "boolean"
          },
          "id": {
            "format": "uuid4",
            "title": "Id",
            "type": "string"
          },
          "origin": {
            "enum": [
              "live",
              "bundled_demo"
            ],
            "title": "Origin",
            "type": "string"
          },
          "paper_limit": {
            "title": "Paper Limit",
            "type": "integer"
          },
          "phase": {
            "enum": [
              "queued",
              "initializing",
              "search",
              "acquisition",
              "chunking",
              "retrieval",
              "analysis",
              "synthesis",
              "citation_check",
              "publishing",
              "terminal"
            ],
            "title": "Phase",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/RunProgress"
          },
          "question": {
            "title": "Question",
            "type": "string"
          },
          "retrieval": {
            "$ref": "#/components/schemas/RetrievalRequest"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "enum": [
              "queued",
              "running",
              "completed",
              "completed_with_degradation",
              "failed",
              "interrupted"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "artifact_run_id",
          "origin",
          "status",
          "phase",
          "question",
          "paper_limit",
          "content_mode",
          "retrieval",
          "progress",
          "created_at",
          "started_at",
          "finished_at",
          "has_report",
          "demo"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "TechScoutCreateRunRequest": {
        "additionalProperties": false,
        "properties": {
          "budget_constraints": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Budget Constraints",
            "type": "array"
          },
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/TechScoutCandidateInput"
            },
            "maxItems": 3,
            "title": "Candidates",
            "type": "array"
          },
          "current_stack": {
            "items": {
              "type": "string"
            },
            "maxItems": 20,
            "title": "Current Stack",
            "type": "array"
          },
          "environment": {
            "$ref": "#/components/schemas/TechScoutEnvironmentRequest"
          },
          "hard_constraints": {
            "items": {
              "type": "string"
            },
            "maxItems": 5,
            "minItems": 1,
            "title": "Hard Constraints",
            "type": "array"
          },
          "license_requirements": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "License Requirements",
            "type": "array"
          },
          "mode": {
            "default": "fast",
            "enum": [
              "fast",
              "verified"
            ],
            "title": "Mode",
            "type": "string"
          },
          "performance_requirements": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Performance Requirements",
            "type": "array"
          },
          "preferences": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Preferences",
            "type": "array"
          },
          "project_context": {
            "maxLength": 2000,
            "minLength": 3,
            "title": "Project Context",
            "type": "string"
          },
          "question": {
            "maxLength": 1000,
            "minLength": 3,
            "title": "Question",
            "type": "string"
          },
          "security_requirements": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Security Requirements",
            "type": "array"
          },
          "team_capabilities": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Team Capabilities",
            "type": "array"
          },
          "use_cases": {
            "items": {
              "type": "string"
            },
            "maxItems": 12,
            "title": "Use Cases",
            "type": "array"
          }
        },
        "required": [
          "question",
          "project_context",
          "environment",
          "hard_constraints"
        ],
        "title": "TechScoutCreateRunRequest",
        "type": "object"
      },
...
      "TechScoutRunList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TechScoutRunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "TechScoutRunList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MOMO TechScout Web API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/runs": {
      "get": {
        "operationId": "list_runs_api_v1_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_v1_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid4",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v2/runs": {
      "get": {
        "operationId": "list_runs_api_v2_runs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechScoutRunList"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "List Runs",
        "tags": [
          "techscout-runs"
        ]
      },
      "post": {
        "operationId": "create_run_api_v2_runs_post",
        "parameters": [
          {
            "in": "header",
            "name": "idempotency-key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TechScoutCreateRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechScoutRunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Run",
        "tags": [
          "techscout-runs"
        ]
      }
    },
    "/api/v2/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v2_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid4",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TechScoutRunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Get Run",
        "tags": [
          "techscout-runs"
        ]
      }
    },
...
  }
}